# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type VocabularyStateSettings = { [key: string]: unknown };

export type VocabularyStateDiscoveryHistoryItem = { [key: string]: unknown };

export interface VocabularyState {
  settings: VocabularyStateSettings;
  progress: unknown[];
  sessions: unknown[];
  bookmarks: string[];
  discoveryHistory?: VocabularyStateDiscoveryHistoryItem[];
}

export interface VocabularyStateInput {
  state: VocabularyState;
}

export interface VocabularyStateResponse {
  state: VocabularyState | null;
  updatedAt: string | null;
}

export interface WordRecommendation {
  wordKey: string;
  word: string;
  field: string;
  count: number;
  recommendedByMe: boolean;
}

export type WordRecommendationInputField = typeof WordRecommendationInputField[keyof typeof WordRecommendationInputField];


export const WordRecommendationInputField = {
  Economics: 'Economics',
  Science: 'Science',
  'Politics_&_Society': 'Politics & Society',
  'Literature_&_Arts': 'Literature & Arts',
  Psychology: 'Psychology',
  'Personality_&_Human_Behavior': 'Personality & Human Behavior',
  'Medicine_&_Healthcare': 'Medicine & Healthcare',
  Technology: 'Technology',
  Everyday_Communication: 'Everyday Communication',
  SAT_Vocabulary: 'SAT Vocabulary',
  Expert_Vocabulary: 'Expert Vocabulary',
} as const;

export type WordRecommendationInputSource = typeof WordRecommendationInputSource[keyof typeof WordRecommendationInputSource];


export const WordRecommendationInputSource = {
  library: 'library',
  search: 'search',
} as const;

export interface WordRecommendationInput {
  /**
     * @minLength 1
     * @maxLength 120
     */
  wordKey: string;
  /**
     * @minLength 1
     * @maxLength 80
     */
  word: string;
  field: WordRecommendationInputField;
  source: WordRecommendationInputSource;
}

export interface WordRecommendationResponse {
  recommendation: WordRecommendation;
}

export interface WordRecommendationList {
  recommendations: WordRecommendation[];
}

export type DiscoverDictionaryEntryField = typeof DiscoverDictionaryEntryField[keyof typeof DiscoverDictionaryEntryField];


export const DiscoverDictionaryEntryField = {
  Economics: 'Economics',
  Science: 'Science',
  'Politics_&_Society': 'Politics & Society',
  'Literature_&_Arts': 'Literature & Arts',
  Psychology: 'Psychology',
  'Personality_&_Human_Behavior': 'Personality & Human Behavior',
  'Medicine_&_Healthcare': 'Medicine & Healthcare',
  Technology: 'Technology',
  Everyday_Communication: 'Everyday Communication',
  SAT_Vocabulary: 'SAT Vocabulary',
  Expert_Vocabulary: 'Expert Vocabulary',
} as const;

export interface DiscoverDictionaryEntry {
  word: string;
  pronunciation: string;
  partOfSpeech: string;
  definition: string;
  field: DiscoverDictionaryEntryField;
  fieldExplanation: string;
  context: string;
  synonyms: string[];
  antonyms: string[];
  wordFamily: string;
  etymology: string;
}

export type LookupDiscoverDictionaryParams = {
/**
 * @minLength 2
 * @maxLength 80
 */
word: string;
};