# lib/api-zod/src/generated/types/discoverDictionaryEntry.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DiscoverDictionaryEntryField } from './discoverDictionaryEntryField';

export interface DiscoverDictionaryEntry {
  word: string;
  pronunciation: string;
  partOfSpeech: string;
  definition: string;
  field: DiscoverDictionaryEntryField;
  fieldExplanation: string;
  context: string;
  synonyms: string[];
  antonyms: string[];
  wordFamily: string;
  etymology: string;
}