# lib/api-zod/src/generated/types/vocabularyState.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { VocabularyStateDiscoveryHistoryItem } from './vocabularyStateDiscoveryHistoryItem';
import type { VocabularyStateSettings } from './vocabularyStateSettings';

export interface VocabularyState {
  settings: VocabularyStateSettings;
  progress: unknown[];
  sessions: unknown[];
  bookmarks: string[];
  discoveryHistory?: VocabularyStateDiscoveryHistoryItem[];
}