# lib/api-zod/src/generated/types/vocabularyStateInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { VocabularyState } from './vocabularyState';

export interface VocabularyStateInput {
  state: VocabularyState;
}