# lib/api-zod/src/generated/types/vocabularyStateResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { VocabularyState } from './vocabularyState';

export interface VocabularyStateResponse {
  state: VocabularyState | null;
  updatedAt: Date | null;
}