# lib/api-zod/src/generated/types/wordRecommendation.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface WordRecommendation {
  wordKey: string;
  word: string;
  field: string;
  count: number;
  recommendedByMe: boolean;
}