# lib/api-zod/src/generated/types/wordRecommendationInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { WordRecommendationInputField } from './wordRecommendationInputField';
import type { WordRecommendationInputSource } from './wordRecommendationInputSource';

export interface WordRecommendationInput {
  /**
     * @minLength 1
     * @maxLength 120
     */
  wordKey: string;
  /**
     * @minLength 1
     * @maxLength 80
     */
  word: string;
  field: WordRecommendationInputField;
  source: WordRecommendationInputSource;
}