# lib/api-zod/src/generated/types/wordRecommendationList.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { WordRecommendation } from './wordRecommendation';

export interface WordRecommendationList {
  recommendations: WordRecommendation[];
}