# lib/api-zod/src/generated/types/wordRecommendationResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { WordRecommendation } from './wordRecommendation';

export interface WordRecommendationResponse {
  recommendation: WordRecommendation;
}